ACIF104-Fase2/requirements.txt
pandas
numpy
scikit-learn
tensorflow
matplotlib
seaborn
fastapi
uvicorn
joblib
shap
python-multipart
xgboost
imbalanced-learn
scipy
scikit-learn-extra
pyyaml